ref: Move propagator out of integrations
