--- sentry_sdk/integrations/opentelemetry/__init__.py
@@ -1,5 +1,0 @@
-from sentry_sdk.integrations.opentelemetry.propagator import SentryPropagator
-
-__all__ = [
-    "SentryPropagator",
-]
