+++ sentry_sdk/integrations/opentelemetry/integration.py
@@ -5,7 +5,7 @@ removed at any time without prior notice.
 """
 
 from sentry_sdk.integrations import DidNotEnable, Integration
-from sentry_sdk.integrations.opentelemetry.propagator import SentryPropagator
+from sentry_sdk.propagator import SentryPropagator
 from sentry_sdk.sampler import SentrySampler
 from sentry_sdk.span_processor import (
     SentrySpanProcessor,
