+++ setup.py
@@ -85,9 +85,7 @@ setup(
         "unleash": ["UnleashClient>=6.0.1"],
     },
     entry_points={
-        "opentelemetry_propagator": [
-            "sentry=sentry_sdk.integrations.opentelemetry:SentryPropagator"
-        ]
+        "opentelemetry_propagator": ["sentry=sentry_sdk.propagator:SentryPropagator"]
     },
     classifiers=[
         "Development Status :: 5 - Production/Stable",
